fix: duplicate instance template
